Remove extra hooks for onChange and onPaste
