+++ src/client/pages/OriginalDataPoint/components/hooks/useUpdateOriginalData.ts
@@ -1,38 +1,21 @@
 import { useCallback } from 'react'
 
 import { CountryIso } from 'meta/area'
-import { ODPNationalClass, ODPs } from 'meta/assessment'
+import { OriginalDataPoint } from 'meta/assessment'
 
 import { useAppDispatch } from 'client/store'
-import { OriginalDataPointActions, useOriginalDataPoint } from 'client/store/ui/originalDataPoint'
+import { OriginalDataPointActions } from 'client/store/ui/originalDataPoint'
 import { useSectionRouteParams } from 'client/hooks/useRouteParams'
-import { Sanitizer } from 'client/utils/sanitizer'
 
-type Props = {
-  field: keyof ODPNationalClass
-  value: string
-  index: number
-}
-
-export const useUpdateOriginalData = (): ((props: Props) => void) => {
+export const useUpdateOriginalData = (): ((originalDataPoint: OriginalDataPoint) => void) => {
   const { assessmentName, cycleName, countryIso, sectionName } = useSectionRouteParams()
   const dispatch = useAppDispatch()
-  const originalDataPoint = useOriginalDataPoint()
 
   return useCallback(
-    (props: Props) => {
-      const { field, value, index } = props
-      const nationalClass = originalDataPoint.nationalClasses[index]
-      const prevValue = nationalClass[field] as string
-
+    (originalDataPoint: OriginalDataPoint) => {
       dispatch(
         OriginalDataPointActions.updateOriginalDataPointOriginalData({
-          originalDataPoint: ODPs.updateNationalClass({
-            odp: originalDataPoint,
-            index,
-            field,
-            value: Sanitizer.acceptNextDecimal(value, prevValue),
-          }),
+          originalDataPoint,
           assessmentName,
           cycleName,
           countryIso: countryIso as CountryIso,
@@ -40,6 +23,6 @@ export const useUpdateOriginalData = (): ((props: Props) => void) => {
         })
       )
     },
-    [assessmentName, countryIso, cycleName, dispatch, originalDataPoint, sectionName]
+    [assessmentName, countryIso, cycleName, dispatch, sectionName]
   )
 }
